Added additional unit tests
